get rid of the whereClauseForNameInArray funny business
